Integrate button at the bottom

+++ Kolumbus/Kolumbus/JHInputTableViewController.h
@@ -13,9 +13,9 @@
 @class JHButton;
 @interface JHInputTableViewController : UITableViewController
 @property (strong,nonatomic) CLLocation* userLocation;
-@property (nonatomic) NSInteger budget;
-@property (nonatomic) NSInteger intensity;
-@property (nonatomic) NSInteger visitCount;
+@property (nonatomic) int budget;
+@property (nonatomic) int intensity;
+@property (nonatomic) int visitCount;
 
 @property (nonatomic,strong) NSDate *startDate; //NYI
 @property (nonatomic,strong) NSDate *endDate;  //NYI
